[oauth] Move back OAuth2 specific classes to the oauth2 package
