+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/security/SaslAuthenticator.java
@@ -22,8 +22,8 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.GenericFutureListener;
 import io.streamnative.pulsar.handlers.kop.KafkaServiceConfiguration;
-import io.streamnative.pulsar.handlers.kop.security.auth.KopOAuthBearerSaslServer;
-import io.streamnative.pulsar.handlers.kop.security.auth.KopOAuthBearerUnsecuredValidatorCallbackHandler;
+import io.streamnative.pulsar.handlers.kop.security.oauth.KopOAuthBearerSaslServer;
+import io.streamnative.pulsar.handlers.kop.security.oauth.KopOAuthBearerUnsecuredValidatorCallbackHandler;
 import io.streamnative.pulsar.handlers.kop.utils.KafkaResponseUtils;
 import java.nio.ByteBuffer;
 import java.util.Collections;
