+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/security/oauth/KopOAuthBearerUnsecuredValidatorCallbackHandler.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.kop.security.auth;
+package io.streamnative.pulsar.handlers.kop.security.oauth;
 
 import java.io.IOException;
 import java.util.Collections;
