+++ kafka-impl/src/test/java/com/datastax/oss/kafka/oauth/ServerConfigTest.java
@@ -13,7 +13,7 @@
  */
 package com.datastax.oss.kafka.oauth;
 
-import io.streamnative.pulsar.handlers.kop.security.auth.ServerConfig;
+import io.streamnative.pulsar.handlers.kop.security.oauth.ServerConfig;
 import java.util.HashMap;
 import java.util.Map;
 import org.testng.Assert;
