+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/CustomOAuthBearerCallbackHandlerTest.java
@@ -19,8 +19,8 @@ import static org.testng.Assert.assertTrue;
 
 import com.google.common.collect.Sets;
 import io.jsonwebtoken.SignatureAlgorithm;
-import io.streamnative.pulsar.handlers.kop.security.auth.KopOAuthBearerToken;
-import io.streamnative.pulsar.handlers.kop.security.auth.KopOAuthBearerValidatorCallback;
+import io.streamnative.pulsar.handlers.kop.security.oauth.KopOAuthBearerToken;
+import io.streamnative.pulsar.handlers.kop.security.oauth.KopOAuthBearerValidatorCallback;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
