+++ tests/src/test/java/io/streamnative/pulsar/handlers/kop/TransactionWithOAuthBearerAuthTest.java
@@ -15,7 +15,7 @@ package io.streamnative.pulsar.handlers.kop;
 
 import com.datastax.oss.kafka.oauth.OauthLoginCallbackHandler;
 import com.google.common.collect.Sets;
-import io.streamnative.pulsar.handlers.kop.security.auth.OauthValidatorCallbackHandler;
+import io.streamnative.pulsar.handlers.kop.security.oauth.OauthValidatorCallbackHandler;
 import java.net.URL;
 import java.util.Properties;
 import lombok.extern.slf4j.Slf4j;
